lib: move CommidId/ChangeId display tests to backend module
