+++ lib/src/object_id.rs
@@ -280,19 +280,8 @@ impl<T: Clone> PrefixResolution<T> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::backend::ChangeId;
-    use crate::backend::CommitId;
 
-    #[test]
-    fn test_display_object_id() {
-        let commit_id = CommitId::from_hex("deadbeef0123");
-        assert_eq!(format!("{commit_id}"), "deadbeef0123");
-        assert_eq!(format!("{commit_id:.6}"), "deadbe");
-
-        let change_id = ChangeId::from_hex("deadbeef0123");
-        assert_eq!(format!("{change_id}"), "mlpmollkzyxw");
-        assert_eq!(format!("{change_id:.6}"), "mlpmol");
-    }
+    id_type!(SomeId { hex() });
 
     #[test]
     fn test_hex_prefix_prefixes() {
@@ -317,7 +306,7 @@ mod tests {
 
     #[test]
     fn test_hex_prefix_matches() {
-        let id = CommitId::from_hex("1234");
+        let id = SomeId::from_hex("1234");
 
         assert!(HexPrefix::try_from_hex("").unwrap().matches(&id));
         assert!(HexPrefix::try_from_hex("1").unwrap().matches(&id));
